+++ src/main/java/trackers/demo/auth/MemberOnly.java
@@ -1,0 +1,11 @@
+package trackers.demo.auth;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface MemberOnly {
+}
