+++ src/main/java/trackers/demo/auth/domain/Authority.java
@@ -1,0 +1,8 @@
+package trackers.demo.auth.domain;
+
+public enum Authority {
+
+    USER,
+
+    ADMIN
+}
