--- src/main/java/trackers/demo/domain/project/domain/Subject.java
@@ -1,21 +1,0 @@
-package trackers.demo.domain.project.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import trackers.demo.global.common.BaseEntity;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Subject extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "subject_id")
-    private Long id;
-
-    @Column(length = 10, nullable = false)
-    private String subjectTitle;
-}
