--- src/main/java/trackers/demo/global/auth/controller/MainController.java
@@ -1,23 +1,0 @@
-package trackers.demo.global.auth.controller;
-
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Collection;
-import java.util.Iterator;
-
-@RestController
-public class MainController {
-
-    @GetMapping("/oauth2/home")
-    public String mainAPI(){
-        return "main route";
-    }
-
-
-}
