--- src/main/java/trackers/demo/global/auth/jwt/JWTFilter.java
@@ -1,87 +1,0 @@
-package trackers.demo.global.auth.jwt;
-
-import io.jsonwebtoken.ExpiredJwtException;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.AllArgsConstructor;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.filter.OncePerRequestFilter;
-import trackers.demo.global.auth.oauth.dto.CustomOAuth2User;
-import trackers.demo.global.auth.oauth.dto.UserDTO;
-
-import java.io.IOException;
-import java.io.PrintWriter;
-
-@RequiredArgsConstructor
-public class JWTFilter extends OncePerRequestFilter {   // 필터는 한 번만 요청되면 됨
-
-    private final JWTUtil jwtUtil;
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-
-        // 헤더에서 access 키에 담긴 토큰을 꺼냄
-        String accessToken = request.getHeader("access");
-
-        // 토큰이 없다면 다음 필터로 넘김
-        if(accessToken == null){
-            filterChain.doFilter(request, response);
-            return;
-        }
-
-        // 토큰 만료 여부 확인, 만료시 다음 필터로 넘기지 않음
-        try {
-            jwtUtil.isExpired(accessToken);
-        } catch (ExpiredJwtException e) {
-
-            //response body
-            PrintWriter writer = response.getWriter();
-            writer.print("access token expired");
-
-            //response status code
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            return;
-        }
-
-        // 토큰이 access인지 확인 (발급시 페이로드에 명시), 다음 필터로 넘기지 않음
-        String category = jwtUtil.getCategory(accessToken);
-
-        if (!category.equals("access")) {
-
-            //response body
-            PrintWriter writer = response.getWriter();
-            writer.print("invalid access token");
-
-            //response status code
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            return;
-        }
-
-        //토큰에서 authKey와 role 획득
-        String authKey = jwtUtil.getAuthkey(accessToken);
-        String role = jwtUtil.getRole(accessToken);
-
-        //UserDTO를 생성하여 빌드
-        UserDTO userDTO = UserDTO.builder()
-                .authKey(authKey)
-                .role(role)
-                .build();
-
-        //UserDetails에 회원 정보 객체 담기
-        CustomOAuth2User customOAuth2User = new CustomOAuth2User(userDTO);
-
-        //스프링 시큐리티 인증 토큰 생성
-        Authentication authToken = new UsernamePasswordAuthenticationToken(customOAuth2User, null, customOAuth2User.getAuthorities());
-
-        //세션에 사용자 등록 -> 요청에 대해서 로그인된 상태로 변경
-        SecurityContextHolder.getContext().setAuthentication(authToken);
-
-        filterChain.doFilter(request, response);
-
-    }
-}
