--- src/main/java/trackers/demo/global/auth/oauth/dto/GoogleResponse.java
@@ -1,33 +1,0 @@
-package trackers.demo.global.auth.oauth.dto;
-
-import java.util.Map;
-
-public class GoogleResponse implements OAuth2Response{
-
-    private final Map<String, Object> attribute;
-
-    public GoogleResponse(Map<String, Object> attribute) {
-        this.attribute = attribute;
-    }
-
-
-    @Override
-    public String getProvider() {
-        return "google";
-    }
-
-    @Override
-    public String getProviderId() {
-        return attribute.get("sub").toString();
-    }
-
-    @Override
-    public String getEmail() {
-        return attribute.get("email").toString();
-    }
-
-    @Override
-    public String getUsername() {
-        return attribute.get("name").toString();
-    }
-}
