--- src/main/java/trackers/demo/global/auth/oauth/dto/UserDTO.java
@@ -1,15 +1,0 @@
-package trackers.demo.global.auth.oauth.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-public class UserDTO {
-
-    private String role;
-    private String username;
-    private String authKey;
-}
