+++ src/main/java/trackers/demo/login/domain/repository/RefreshTokenRepository.java
@@ -1,0 +1,13 @@
+package trackers.demo.login.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.transaction.annotation.Transactional;
+import trackers.demo.login.domain.RefreshToken;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+
+    Boolean existsByRefresh(String refresh);
+
+    @Transactional
+    void deleteByRefresh(String refresh);
+}
