+++ src/main/java/trackers/demo/login/oauth/dto/UserDTO.java
@@ -1,0 +1,15 @@
+package trackers.demo.login.oauth.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+public class UserDTO {
+
+    private String role;
+    private String username;
+    private String authKey;
+}
