+++ src/main/java/trackers/demo/login/service/Authservice.java
@@ -1,0 +1,102 @@
+package trackers.demo.login.service;
+
+import io.jsonwebtoken.ExpiredJwtException;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
+import trackers.demo.login.jwt.JWTUtil;
+import trackers.demo.login.domain.RefreshToken;
+import trackers.demo.login.domain.repository.RefreshTokenRepository;
+
+import java.util.Date;
+
+@Service
+@RequiredArgsConstructor
+public class Authservice {
+
+    private final JWTUtil jwtUtil;
+    private final RefreshTokenRepository refreshTokenRepository;
+
+    public ResponseEntity<?> reissue(HttpServletRequest request, HttpServletResponse response) {
+
+        // refresh 토큰 받기
+        String refresh = null;
+        Cookie[] cookies = request.getCookies();
+
+        for (Cookie cookie : cookies) {
+            if (cookie.getName().equals("refresh")) {
+                refresh = cookie.getValue();
+            }
+        }
+
+        if (refresh == null) {
+            return new ResponseEntity<>("refresh token null", HttpStatus.BAD_REQUEST);
+        }
+
+        // refresh 만료 여부 확인
+        try {
+            jwtUtil.isExpired(refresh);
+        } catch (ExpiredJwtException e) {
+            return new ResponseEntity<>("refresh token expired", HttpStatus.BAD_REQUEST);
+        }
+
+        // 토큰이 refresh인지 여부 확인
+        String category = jwtUtil.getCategory(refresh);
+
+        if (!category.equals("refresh")) {
+            return new ResponseEntity<>("invalid refresh token", HttpStatus.BAD_REQUEST);
+        }
+
+        // DB에 refresh 토큰이 저장되어 있는지 여부 확인
+        Boolean isExist = refreshTokenRepository.existsByRefresh(refresh);
+        if (!isExist) {
+            //response body
+            return new ResponseEntity<>("invalid refresh token", HttpStatus.BAD_REQUEST);
+        }
+
+        // 새로운 JWT 발급
+        String authKey = jwtUtil.getAuthkey(refresh);
+        String role = jwtUtil.getRole(refresh);
+
+        String newAccess = jwtUtil.createJwt("access", authKey, role, 600000L);
+        String newRefresh = jwtUtil.createJwt("refresh", authKey, role, 86400000L);
+
+        //Refresh 토큰 저장 DB에 기존의 Refresh 토큰 삭제 후 새 Refresh 토큰 저장
+        refreshTokenRepository.deleteByRefresh(refresh);
+        addRefreshEntity(authKey, newRefresh, 86400000L);
+
+        response.addHeader("access", newAccess);
+        response.addCookie(createCookie("refresh", newRefresh));
+
+        return new ResponseEntity<>(HttpStatus.OK);
+
+    }
+
+    private void addRefreshEntity(String authkey, String refresh, Long expiredMs) {
+
+        Date date = new Date(System.currentTimeMillis() + expiredMs);
+
+        RefreshToken refreshTokenEntity = RefreshToken.builder()
+                .authkey(authkey)
+                .refresh(refresh)
+                .expiration(date.toString())
+                .build();
+
+        refreshTokenRepository.save(refreshTokenEntity);
+    }
+
+    private Cookie createCookie(String key, String value) {
+
+        Cookie cookie = new Cookie(key, value);
+        cookie.setMaxAge(24*60*60);
+        //cookie.setSecure(true);
+        //cookie.setPath("/");
+        cookie.setHttpOnly(true);
+
+        return cookie;
+    }
+}
