+++ src/main/java/trackers/demo/project/presentation/ProjectController.java
@@ -1,0 +1,14 @@
+package trackers.demo.project.presentation;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/project")
+public class ProjectController {
+
+//    @PostMapping("/first")
+}
