+++ src/main/java/trackers/demo/user/domain/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package trackers.demo.user.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import trackers.demo.user.domain.User;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Boolean existsByAuthkey(String authKey);
+    User findByAuthkey(String authkey);
+    Boolean existsByUsername(String username);
+    User findByUsername(String username);
+}
